perf(decoder): shrink FLACDecoder ~100 bytes (int16_t coefs + lazy metadata)
Narrow LpcState::coefs from int32_t[32] to int16_t[32] (64 B saved): FLAC qlp coefficient precision is bounded to <=15 bits per RFC 9639 §9.2.6, so values fit in int16_t with no loss. Updates the Xtensa LPC assembly to load coefs via l16si (halved offsets, addx4->addx2 in the generic path), the C++ fallback signatures, FIXED_COEFFICIENTS tables, and the test_lpc_asm reference vectors + generator.

Defer max_metadata_sizes_ (32 B inline -> 4 B pointer) and metadata_blocks_ to heap allocation; both are lazily populated only when the opt-in set_max_metadata_size() / metadata-storage features are used. Repack HeaderParseState (narrow type to uint8_t, drop redundant data_capacity). Also adds an in-loop bound check on the wasted-bits shift so pathological input is rejected immediately. Hardware-verified on ESP32-S3: zero perf regression vs baseline (917.6 ms full-frame, 32.7x real-time), 192/192 LPC assembly test vectors pass.

+++ src/lpc.h
@@ -34,11 +34,12 @@ namespace micro_flac {
 /// in-place)
 /// @param num_of_samples Total number of samples in the buffer
 /// @param bits_per_sample Bit depth of the audio samples
-/// @param coefs Pointer to array of LPC coefficients
+/// @param coefs Pointer to array of LPC coefficients. Stored as int16_t because the FLAC
+///        spec bounds precision to <= 15 bits (RFC 9639 §9.2.6, values in [-16384, 16383]).
 /// @param order Number of LPC coefficients (predictor order)
 /// @param shift Right shift amount to apply after prediction
 void restore_lpc(int32_t* sub_frame_buffer, size_t num_of_samples, uint32_t bits_per_sample,
-                 const int32_t* coefs, uint32_t order, int32_t shift);
+                 const int16_t* coefs, uint32_t order, int32_t shift);
 
 /// @brief Restore linear prediction for 33-bit MID_SIDE side channel
 ///
@@ -49,10 +50,10 @@ void restore_lpc(int32_t* sub_frame_buffer, size_t num_of_samples, uint32_t bits
 /// @param sub_frame_buffer Buffer containing int64_t warm-up samples followed by int64_t residuals
 /// @param num_of_samples Total number of samples in the buffer
 /// @param bits_per_sample Accepted for API symmetry with int32_t overload (unused)
-/// @param coefs Pointer to array of LPC coefficients (int32_t)
+/// @param coefs Pointer to array of LPC coefficients (see int32_t overload for type rationale)
 /// @param order Number of LPC coefficients (predictor order)
 /// @param shift Right shift amount to apply after prediction
 void restore_lpc(int64_t* sub_frame_buffer, size_t num_of_samples, uint32_t bits_per_sample,
-                 const int32_t* coefs, uint32_t order, int32_t shift);
+                 const int16_t* coefs, uint32_t order, int32_t shift);
 
 }  // namespace micro_flac
